% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_nms.R
\name{check_col_nms}
\alias{check_col_nms}
\title{Check that column names in the template and metadata files match.}
\usage{
check_col_nms(tableName, datIn, err, ...)
}
\arguments{
\item{tableName}{specific table name to check}

\item{datIn}{input data to check}

\item{err}{error counter; defaults to 0}
}
\value{
error counter "err"
}
\description{
Check that column names in the template and metadata files match.
}
\note{
This is typically only called from \code{\link{compile_core}}.
}
\keyword{internal}
