% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sweddie_core.R
\name{check_sweddie_core}
\alias{check_sweddie_core}
\title{Check whether an object contains SWEDDIE core tables}
\usage{
check_sweddie_core(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
logical
}
\description{
A valid SWEDDIE object is a list with the following elements: experiment, site, plot
}
\keyword{internal}
