% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_core.R
\name{compile_core}
\alias{compile_core}
\title{Build core SWEDDIE database}
\usage{
compile_core(
  DIR = "~/sweddie_db",
  write_report = FALSE,
  verbose = FALSE,
  EOL_err = FALSE
)
}
\arguments{
\item{DIR}{parent directory in which SWEDDIE 'database' directory is stored}

\item{write_report}{logical; should report of build be written to a file?}

\item{verbose}{logical}

\item{EOL_err}{set to true if you encounter an end of line (EOL) error}
}
\value{
list
}
\description{
returns SWEDDIE core database object with experiment, site, and plot tables
}
