% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_meta.R
\name{compile_meta}
\alias{compile_meta}
\title{Compile SWEDDIE metadata}
\usage{
compile_meta(
  DIR = "~/eco-warm/data/sweddie",
  expName,
  verbose = TRUE,
  write_report = FALSE
)
}
\arguments{
\item{DIR}{parent directory in which SWEDDIE 'database' directory is stored}

\item{expName}{name of experiment to retrieve metadata from; must match standard names}

\item{verbose}{logical}

\item{write_report}{logical; should report be written to a file?}
}
\value{
list
}
\description{
returns SWEDDIE metadata object
}
