% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_meta_one.R
\name{compile_meta_one}
\alias{compile_meta_one}
\title{Compile SWEDDIE metadata for single experiment}
\usage{
compile_meta_one(DIR, expName)
}
\arguments{
\item{DIR}{local directory for SWEDDIE database}

\item{expName}{name of experiment to retrieve metadata from; must match standard names}

\item{write_report}{logical; should report be written to a file?}

\item{verbose}{logical}
}
\value{
list
}
\description{
returns SWEDDIE metadata object
}
