% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_sweddie.R
\name{compile_sweddie}
\alias{compile_sweddie}
\title{Compile SWEDDIE data}
\usage{
compile_sweddie(DIR = "~/sweddie_db", verbose = TRUE)
}
\arguments{
\item{DIR}{local directory for SWEDDIE files}

\item{verbose}{should function progress be passed to console?}
}
\value{
list
}
\description{
General function for compiling SWEDDIE data. The resulting list object will reflect the directory structure, i.e., each recursive directory is treated
as a subsequent list element and each (CSV or compressed CSV) file within a directory as a data frame
element within that list. Be wary of compiling all available variables as this will be very slow.
}
