% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_add.R
\name{cr_add}
\alias{cr_add}
\title{Add carriage return to .csv files}
\usage{
cr_add(dataDir, ..)
}
\arguments{
\item{dataDir}{directory containing .csv files}
}
\value{
modified .csv files
}
\description{
Adds a carriage return (end of line character) to .csv files. Note that this is recursive by default. This behavior can be changed by setting "recursive = FALSE".
}
\keyword{internal}
