% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_helper.R
\name{dd_helper}
\alias{dd_helper}
\title{Helper for creating data dictionary files}
\usage{
dd_helper(
  expName,
  dataName,
  DATA_DIR = NULL,
  META_DIR = NULL,
  write_out = TRUE
)
}
\arguments{
\item{expName}{name of experiment}

\item{dataName}{input data to check}

\item{DATA_DIR}{directory to search for .csv file matching supplied dataName}

\item{META_DIR}{directory to search for .csv file of matching data dictionary file}

\item{write_out}{should function write the template file to disk?}
}
\value{
if write_out = FALSE, data frame of dd template
}
\description{
Helper for creating data dictionary files
}
\details{
this function creates a data dictionary template file for the supplied input data
}
