% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package load hook}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{character; path to the library directory.}

\item{pkgname}{character; name of the package.}
}
\value{
NULL (called for side effects only).
}
\description{
Initialize package-level options and internal state when the SWEDDIE package
is loaded. This function sets default logging behavior and prepares internal
environments required by other package functions.
}
\details{
This function is called automatically by R when the package namespace is
loaded and should not be invoked directly.
}
\keyword{internal}
