% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_plot.R
\name{filPltTbl.fx}
\alias{filPltTbl.fx}
\title{Merge plot table for SWEDDIE database object}
\usage{
filPltTbl.fx(db, treatment = "warming only", plt_cols = NULL)
}
\arguments{
\item{db}{SWEDDIE database object}

\item{treatment}{Enter treatment name. Options are "warming only" (default) or "all".}

\item{plt_cols}{Specify columns to fill from plot table}
}
\description{
Merge plot table for SWEDDIE database object
}
\details{
Function for merging plot table values with data tables in a SWEDDIE database object. Optional filtering for specific treatments.
}
