% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CSV_nms.R
\name{get_CSV_nms}
\alias{get_CSV_nms}
\title{Get csv names}
\usage{
get_CSV_nms(path)
}
\arguments{
\item{path}{path to CSV or GZ file}
}
\value{
names of csv file columns
}
\description{
allows name checking without loading whole file
}
\keyword{internal}
