% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FLMD_dd.R
\name{get_FLMD_dd}
\alias{get_FLMD_dd}
\title{Read template files}
\usage{
get_FLMD_dd(exp_names = NULL)
}
\arguments{
\item{exp_names}{list of directories containing SWEDDIE data}
}
\value{
A list with core template tables or core meta data
}
\description{
Wrapper function to run fx \code{\link{read_core_templates}} on all files in list of directories specified by "exp_names"
}
\details{
Each directory in the list 'exp_names' must contain the directories named "input_data" and "meta", per \code{\link{read_core_templates}} specification
}
