% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varNames.R
\name{get_varNames}
\alias{get_varNames}
\title{Helper for retrieving unique variable names}
\usage{
get_varNames(meta = NULL)
}
\arguments{
\item{meta}{object returned by compile_meta}
}
\description{
Helper for retrieving unique variable names
}
