% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestDat.R
\name{ingestDat}
\alias{ingestDat}
\title{Harmonization function for ingesting raw data into SWEDDIE}
\usage{
ingestDat(
  DIR = "~/sweddie_db",
  expName,
  path.dat.csv,
  path.dd.csv,
  append.flmd,
  ...
)
}
\arguments{
\item{DIR}{local directory for SWEDDIE}

\item{expName}{name of experiment}

\item{path.dat.csv}{path to raw data file}

\item{path.dd.csv}{path to data dictionary file of raw data}

\item{append.flmd}{should new rows be added to existing FLMD file?}

\item{...}{used internally for optional arguments passed to function}
}
\description{
Harmonization function for ingesting raw data into SWEDDIE
}
\details{
interactive function for harmonizing raw data which outputs dat files and their dd files, and optionally updates FLMD
}
