% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_core_templates.R
\name{read_core_templates}
\alias{read_core_templates}
\title{Read core template files}
\usage{
read_core_templates(templateDir, return = "template")
}
\arguments{
\item{templateDir}{core template directory}

\item{return}{should an empty template list be returned ("template") or the metadata file itself?}
}
\value{
A list with core template tables or core meta data
}
\description{
loads
}
