% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_cmp.R
\name{read_csv_cmp}
\alias{read_csv_cmp}
\title{Helper for reading CSV and GZ files}
\usage{
read_csv_cmp(path, ...)
}
\arguments{
\item{path}{path to file}

\item{...}{optional arguments supplied to read.csv}
}
\value{
file
}
\description{
Helper for reading CSV and GZ files
}
\details{
runs 'read.csv' and decompresses GZ files as needed
}
