% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_str.R
\name{sanitize_str}
\alias{sanitize_str}
\title{Sanitize strings to mitigate encoding issues}
\usage{
sanitize_str(string)
}
\arguments{
\item{string}{string to sanitize}
}
\value{
ASCII//TRANSLIT encoded string
}
\description{
Sanitize strings to mitigate encoding issues
}
\details{
Removes formatting (subscripts, superscripts, slashes) and non-standard characters from string input
}
\keyword{internal}
