% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_csv_ext.R
\name{strip_csv_ext}
\alias{strip_csv_ext}
\title{Utility for stripping CSV and GZ extensions from filenames}
\usage{
strip_csv_ext(path)
}
\arguments{
\item{path}{path to file}
}
\description{
Utility for stripping CSV and GZ extensions from filenames
}
\keyword{internal}
