% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweddie_set_logging.R
\name{sweddie_set_logging}
\alias{sweddie_set_logging}
\title{Configure SWEDDIE logging behavior}
\usage{
sweddie_set_logging(verbose = NULL, file = NULL, append = NULL)
}
\arguments{
\item{verbose}{logical; print messages to console?}

\item{file}{character; path to log file ("" = console only)}

\item{append}{logical; append to existing log file?}
}
\description{
Configure SWEDDIE logging behavior
}
