% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_freq.R, R/utils.R
\name{ts_freq}
\alias{ts_freq}
\title{Check observation frequency}
\usage{
ts_freq(
  df,
  dateName = "Date",
  pltName = "plt_name",
  repName = NULL,
  depth = FALSE
)

ts_freq(
  df,
  dateName = "Date",
  pltName = "plt_name",
  repName = NULL,
  depth = FALSE
)
}
\arguments{
\item{df}{data frame}

\item{dateName}{name of column containing observation dates}

\item{pltName}{name of column with plot identifiers}

\item{depth}{are data depth resolved? If so, data must contain a column called "depth"}
}
\value{
list

list
}
\description{
returns list of median interval in seconds, days, inferred sampling frequency, etc.

returns list of median interval in seconds, days, inferred sampling frequency, etc.
}
\keyword{internal}
