% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_all.R
\name{compile_all}
\alias{compile_all}
\title{Compile all SWEDDIE data}
\usage{
compile_all(
  DIR = "~/sweddie_db/sweddie",
  pattern = "\\\\.csv(\\\\.gz)?$",
  read_fun = read_csv_cmp,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{DIR}{local directory for SWEDDIE data files (default: "~/sweddie_db/sweddie)}

\item{pattern}{regex pattern for files to read (defaults to CSV and GZ)}

\item{read_fun}{function to call for reading files (defaults to
'sweddie::read_csv_comp')}

\item{verbose}{should function progress be passed to console?}

\item{...}{optional arguments to pass to read_fun}
}
\value{
list
}
\description{
General function for compiling all data files within the
specified directory into a list object. Each recursive directory is treated
as a subsequent list element and each file within a directory as a data frame
element within that list. Be wary of running this on the entire SWEDDIE
database as it will be very slow. Observational data are best retrieved with
filtering queries.
}
