% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_core.R
\name{compile_core}
\alias{compile_core}
\title{Build core SWEDDIE database}
\usage{
compile_core(
  DIR = "~/eco-warm/data/sweddie",
  write_report = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{DIR}{parent directory in which SWEDDIE 'database' directory is stored}

\item{write_report}{logical; should report of build be written to a file?}

\item{verbose}{logical}
}
\value{
list
}
\description{
returns SWEDDIE core database object with meta, site, and plot tables
}
