% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_meta.R
\name{compile_meta}
\alias{compile_meta}
\title{Compile SWEDDIE metadata}
\usage{
compile_meta(
  DIR = "~/sweddie_db",
  expName = NULL,
  verbose = TRUE,
  write_report = FALSE,
  EOL_err = FALSE
)
}
\arguments{
\item{DIR}{local directory for SWEDDIE}

\item{expName}{name(s) of experiment(s) to retrieve metadata from; must match standard names}

\item{verbose}{logical}

\item{write_report}{logical; should report be written to a file?}

\item{EOL_err}{set to true if you encounter an end of line (EOL) error}
}
\value{
list
}
\description{
returns SWEDDIE metadata object
}
