% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flmd_helper.R
\name{flmd_helper}
\alias{flmd_helper}
\title{Helper for creating or appending file level metadata file}
\usage{
flmd_helper(
  expName,
  dataFileName,
  dateColName,
  rename = FALSE,
  append = TRUE,
  write_out = TRUE,
  orders = NULL,
  ...
)
}
\arguments{
\item{expName}{name of experiment}

\item{dataFileName}{name of input file (must be *.csv file, do not include extension)}

\item{dateColName}{supply known text string for column in input data containing dates}

\item{rename}{should file name be overwritten, e.g., 'expName_flmd'?}

\item{append}{should new rows be added to existing FLMD file?}

\item{write_out}{should function write out a FLMD file in *.csv format?}

\item{orders}{optionally supply date format, e.g., \%m-\%d-\%Y}

\item{...}{used internally for optional arguments passed to function}
}
\description{
Helper for creating or appending file level metadata file
}
\details{
interactive function for filling out or updating file level metadata (FLMD)
}
